+++ examples/06_datetime_encoder.py
@@ -11,71 +11,93 @@ Blabla
 # --------------
 #
 # We first get the dataset:
+import pandas as pd
 from dirty_cat.datasets import fetch_traffic_violations
 road_safety = fetch_traffic_violations()
 print(road_safety.description)
-print(road_safety)
-print(road_safety.X.columns)
 
 ###############################################################################
-# Now, we retrieve the dirty column to encode:
-dirty_column = 'employee_position_title'
-X_dirty = employee_salaries.X[[dirty_column]]
-print(X_dirty.head(), end='\n\n')
-print(f'Number of dirty entries = {len(X_dirty)}')
+# Now, we select relevant features, choose the target and reduce the number of samples
+data = road_safety.X[["date_of_stop", "time_of_stop", "longitude", "latitude", "gender", "alcohol", "subagency"]]
+y = road_safety.X["accident"] # Whether the traffic violation is linked to an accident
+X = road_safety.X.drop("accident", axis=1)
+# Reduce dataset size for speed
+import numpy as np
+rng = np.random.default_rng(1)
+indices = rng.choice(range(len(y)), 30000)
+X, y = X.iloc[indices], y.iloc[indices]
 
 ###############################################################################
-# Encoding dirty job titles
+# Creating encoders for categorical and datetime features
 # -------------------------
 #
-# We first create an instance of the GapEncoder with n_components=10:
-from dirty_cat import GapEncoder
-enc = GapEncoder(n_components=10, random_state=42)
+from sklearn.preprocessing import OneHotEncoder
+from dirty_cat.datetime_encoder import DatetimeEncoder
+cat_encoder = OneHotEncoder(handle_unknown="ignore")
+datetime_encoder = DatetimeEncoder(add_day_of_the_week=True)
 
 ###############################################################################
-# Then we fit the model on the dirty categorical data and transform it to
-# obtain encoded vectors of size 10:
-X_enc = enc.fit_transform(X_dirty)
-print(f'Shape of encoded vectors = {X_enc.shape}')
+# Creating a column encoder
+from sklearn.compose import make_column_transformer
+datetime_columns = ["date_of_stop", "time_of_stop"]
+numerical_columns = ["longitude", "latitude"]
+categorical_columns = ["gender", "alcohol", "subagency"]
+
+encoder = make_column_transformer((cat_encoder, categorical_columns),
+                                  (datetime_encoder, datetime_columns),
+                                  ("passthrough", numerical_columns),
+                                  remainder="drop")
 
 ###############################################################################
-# Interpreting encoded vectors
+# Transforming the input data
 # ----------------------------
-#
-# The GapEncoder can be understood as a continuous encoding on a set of latent
-# topics estimated from the data. The latent topics are built by
-# capturing combinations of substrings that frequently co-occur, and encoded
-# vectors correspond to their activations.
-# To interpret these latent topics, we select for each of them a few labels
-# from the input data with the highest activations.
-# In the example below we select 3 labels to summarize each topic.
-
-topic_labels = enc.get_feature_names(n_labels=3)
-for k in range(len(topic_labels)):
-    labels = topic_labels[k]
-    print(f'Topic n°{k}: {labels}')
+# We can see that the encoder is working as expected, and new date features are created.
+X_ = encoder.fit_transform(X)
+feature_names = encoder.get_feature_names_out()
+feature_names
 
 ###############################################################################
-# As expected, topics capture labels that frequently co-occur. For instance,
-# the labels *firefighter*, *rescuer*, *rescue* appear together in
-# *Firefigther/Rescuer III*, or *Fire/Rescue Lieutenant*.
-#
-# This enables us to understand the encoding of different samples
-
+# Features importance
+# ----------------------------
+# This allows us to compute the importance of these date features, using the permutation_importance function.
+from sklearn.ensemble import HistGradientBoostingClassifier
+clf = HistGradientBoostingClassifier().fit(X_, y)
+from sklearn.inspection import permutation_importance
+result = permutation_importance(clf, X_, y, n_repeats=10, random_state=0)
+std = result.importances_std
+importances = result.importances_mean
+indices = np.argsort(importances)
+# Sort from least to most
+indices = list(reversed(indices))
 import matplotlib.pyplot as plt
-encoded_labels = enc.transform(X_dirty[:20])
-plt.figure(figsize=(8,10))
-plt.imshow(encoded_labels)
-plt.xlabel('Latent topics', size=12)
-plt.xticks(range(0, 10), labels=topic_labels, rotation=50, ha='right')
-plt.ylabel('Data entries', size=12)
-plt.yticks(range(0, 20), labels=X_dirty[:20].to_numpy().flatten())
-plt.colorbar().set_label(label='Topic activations', size=12)
-plt.tight_layout()
+plt.figure(figsize=(12, 9))
+plt.title("Feature importances")
+n = len(indices)
+labels = np.array(feature_names)[indices]
+plt.barh(range(n), importances[indices], color="b", yerr=std[indices])
+plt.yticks(range(n), labels, size=15)
+plt.tight_layout(pad=1)
 plt.show()
 
-###############################################################################
-# As we can see, each dirty category encodes on a small number of topics,
-# These can thus be reliably used to summarize each topic, which are in
-# effect latent categories captured from the data.
 
+
+###############################################################################
+# One-liner with the SuperVectorizer
+# ----------------------------
+# The DatetimeEncoder is used by default in the SuperVectorizer, which automatically detects datetime features.
+from dirty_cat import SuperVectorizer
+sup_vec = SuperVectorizer()
+X_ = sup_vec.fit_transform(X)
+feature_names = sup_vec.get_feature_names_out()
+feature_names
+
+###############################################################################
+# We can see that the SuperVectorizer is indeed using a DatetimeEncoder for the datetime features.
+sup_vec.transformers_
+
+###############################################################################
+# If we want the day of the week, we can just replace SuperVectorizer's default
+sup_vec = SuperVectorizer(datetime_transformer=DatetimeEncoder(add_day_of_the_week=True))
+X_ = sup_vec.fit_transform(X)
+feature_names = sup_vec.get_feature_names_out()
+feature_names